requirements.txt
# Cleaned requirements.txt by removing invalid local file path entries
annotated-types==0.7.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
boltons==25.0.0
Brotli==1.1.0
build==1.2.2.post1
certifi
cffi
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.3
cssselect2==0.8.0
defusedxml==0.7.1
distro
Django==5.2.2
django-cors-headers==4.7.0
django-debug-toolbar
django-extensions==4.1
django-jalali==7.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-excel==2.5.3
et_xmlfile==2.0.0
fonttools==4.58.2
frozendict==2.4.6
html5lib==1.1
idna==3.10
jalali_core==1.0.0
jdatetime==5.2.0
jsonpatch
jsonpointer==3.0.0
lxml==5.4.0
markdown-it-py==3.0.0
mdurl==0.1.2
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pillow==11.2.1
pip-review==1.3.0
pip-tools==7.4.1
platformdirs==4.3.8
pluggy==1.6.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydyf==0.11.0
Pygments
pyHanko==0.29.0
pyhanko-certvalidator==0.27.0
PyJWT==2.9.0
pypdf==5.6.0
pyphen==0.17.2
pyproject_hooks==1.2.0
PySocks
python-bidi==0.6.6
PyYAML==6.0.2
qrcode==8.2
reportlab==4.4.1
requests
rich==14.0.0
ruamel.yaml==0.18.13
ruamel.yaml.clib
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm
truststore==0.10.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
wheel==0.45.1
xhtml2pdf==0.2.17
xlwt==1.3.0
zopfli==0.2.3.post1
zstandard==0.23.0
gunicorn==23.0.0
dj-database-url==3.0.0
whitenoise[brotli]==6.9.0
fastapi==0.111.1
uvicorn[standard]==0.30.3
django-allauth
django-recaptcha
django-axes
